the creation and the get , set , sellect by Id from the 4 classes is created

+++ src/main/java/com/codeline/sampleProject/Models/Salary.java
@@ -22,13 +22,13 @@ public class Salary extends BaseEntity {
 
     String currency;
 
-    Double overTimeAmount;
+    Integer overTimeAmount;
 
     Double deductions;
 
     Double healthCareContribution;
 
-    Double allowances;
+    Integer allowances;
 
     Double bonus;
